Create audioenf
audio enf. 
on testing, the enf fills to the entrity of the passband. Thus, we need to determine how to nominate an appropriate f_nominal and bandwidth.

Do we require to extract all the enfs from the grids and then take an audio enf for each grid case?
I will raise an issue for this.